MPI in extension and remove old files
